test : fix std::max type mismatch in tape-window test [P1 WS-2 G3]
std::max(uint32_t, int) cannot deduce a common type -- cast the literals.
Caught by the dorei build (the new test target was absent from the local
compile_commands.json, so the local -fsyntax-only pass skipped it).

+++ tests/test-dflash-tape-window.cpp
@@ -20,8 +20,8 @@ static llama_context_ptr make_ctx(const common_params & params, llama_model * mo
     auto cparams = common_context_params_to_llama(params);
     cparams.n_seq_max = 1;
     cparams.n_rs_seq  = 0; // Gate 3 tests the replacement window, not rollback planes.
-    cparams.n_batch   = std::max(cparams.n_batch,  16);
-    cparams.n_ubatch  = std::max(cparams.n_ubatch, 16);
+    cparams.n_batch   = std::max(cparams.n_batch,  (uint32_t) 16);
+    cparams.n_ubatch  = std::max(cparams.n_ubatch, (uint32_t) 16);
     return llama_context_ptr(llama_init_from_model(model, cparams));
 }
